Split methods from models for better integration
src/models/methods/*.js:
+ Add `person_methods.js`, `staff_methods.js`, and `student_methods.js` for method definitions from the referenced models.

`src/models/person.js`:
+ Move methods to associated script in `methods` folder, then import as `methods` object and export it directly
- Delete unused imports and references.

`src/models/staff.js`, and `src/models/student.js`:
* Rename `methods` import from person model to `personMethods` for clarity
+ Move methods to associated script in `methods` folder, import as `methods` object and append them to method declarations
- Delete unused imports and references.
